requirements.txt
allure-pytest
pytest-playwright
pytest-xdist
pytest